change placement of nav
