--- src/tests/helpers/pageRedirect.ts
@@ -1,20 +1,0 @@
-const pageRedirect = [
-  {
-    name: 'Home',
-    url: '/',
-  },
-  {
-    name: 'Residents',
-    url: '/residents',
-  },
-  {
-    name: 'Tenant and Resident Associations',
-    url: '/TRA',
-  },
-  {
-    name: 'Administration',
-    url: '/Administration',
-  },
-];
-
-export default pageRedirect;
